Move reducer tests back to `sites/test/reduccer.js`.
