+++ client/state/sites/test/reducer.js
@@ -38,6 +38,7 @@ import {
 	SITE_PLUGIN_UPDATED,
 	SERIALIZE,
 	DESERIALIZE,
+	EDITOR_TYPE_SET,
 } from 'state/action-types';
 import { withSchemaValidation } from 'state/utils';
 import { useSandbox } from 'test/helpers/use-sinon';
@@ -548,6 +549,68 @@ describe( 'reducer', () => {
 					name: 'WordPress.com Example Blog',
 				},
 			} );
+		} );
+
+		test( "should add the given site's editor property if not present and a valid value is received", () => {
+			const original = deepFreeze( {
+				2916284: {
+					ID: 2916284,
+					name: 'WordPress.com Example Blog',
+				},
+			} );
+			const state = items( original, {
+				type: EDITOR_TYPE_SET,
+				siteId: 2916284,
+				editor: 'gutenberg',
+			} );
+
+			expect( state ).to.eql( {
+				2916284: {
+					ID: 2916284,
+					name: 'WordPress.com Example Blog',
+					editor: 'gutenberg',
+				},
+			} );
+		} );
+
+		test( "should update the given site's editor property if a valid value is received", () => {
+			const original = deepFreeze( {
+				2916284: {
+					ID: 2916284,
+					name: 'WordPress.com Example Blog',
+					editor: 'classic',
+				},
+			} );
+			const state = items( original, {
+				type: EDITOR_TYPE_SET,
+				siteId: 2916284,
+				editor: 'gutenberg',
+			} );
+
+			expect( state ).to.eql( {
+				2916284: {
+					ID: 2916284,
+					name: 'WordPress.com Example Blog',
+					editor: 'gutenberg',
+				},
+			} );
+		} );
+
+		test( "should not update the given site's editor property if an invalid value is received", () => {
+			const original = deepFreeze( {
+				2916284: {
+					ID: 2916284,
+					name: 'WordPress.com Example Blog',
+					editor: 'gutenberg',
+				},
+			} );
+			const state = items( original, {
+				type: EDITOR_TYPE_SET,
+				siteId: 2916284,
+				editor: 'puppers',
+			} );
+
+			expect( state ).to.eql( original );
 		} );
 
 		test( 'should persist state', () => {
